+++ x/erc20/keeper/msg_server.go
@@ -315,18 +315,13 @@ func (k *Keeper) UpdateParams(goCtx context.Context, req *types.MsgUpdateParams)
 	return &types.MsgUpdateParamsResponse{}, nil
 }
 
-// RegisterERC20 implements the gRPC MsgServer interface. After a successful governance vote
-// it updates creates the token pair for an ERC20 contract if the requested authority
-// is the Cosmos SDK governance module account
+// RegisterERC20 implements the gRPC MsgServer interface. Any account can permissionlessly
+// register a native ERC20 contract to map to a Cosmos Coin.
 func (k *Keeper) RegisterERC20(goCtx context.Context, req *types.MsgRegisterERC20) (*types.MsgRegisterERC20Response, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 	// Check if the conversion is globally enabled
 	if !k.IsERC20Enabled(ctx) {
 		return nil, types.ErrERC20Disabled.Wrap("registration is currently disabled by governance")
-	}
-
-	if err := k.validateAuthority(req.Authority); err != nil {
-		return nil, err
 	}
 
 	for _, addr := range req.Erc20Addresses {