+++ x/erc20/types/tx.pb.gw.go
@@ -244,7 +244,7 @@ func RegisterMsgHandlerClient(ctx context.Context, mux *runtime.ServeMux, client
 var (
 	pattern_Msg_ConvertERC20_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3, 2, 4, 2, 5}, []string{"cosmos", "evm", "erc20", "v1", "tx", "convert_erc20"}, "", runtime.AssumeColonVerbOpt(false)))
 
-	pattern_Msg_ConvertCoin_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3, 2, 4}, []string{"os", "erc20", "v1", "tx", "convert_coin"}, "", runtime.AssumeColonVerbOpt(false)))
+	pattern_Msg_ConvertCoin_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1, 2, 2, 2, 3, 2, 4, 2, 5}, []string{"cosmos", "evm", "erc20", "v1", "tx", "convert_coin"}, "", runtime.AssumeColonVerbOpt(false)))
 )
 
 var (